backend/requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
fastapi==0.110.1
fastapi-sessions==0.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
jose==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
plotly==5.20.0
pluggy==1.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.3
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
redis==5.0.3
redit==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
starlette-session==0.4.3
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
